pyproject.toml
[tool.poetry]
name = "bert-squeeze"
version = "0.1.0"
description = ""
authors = ["JulesBelveze <jules.belveze@hotmail.fr>"]

[tool.poetry.dependencies]
python = "^3.7"
transformers = "^4.6.1"
torch = "^1.8.1"
numpy = "^1.20.3"
tabulate = "^0.8.9"
overrides = "^6.1.0"
python-dotenv = "^0.18.0"
datasets = "^1.8.0"
neptune-client = {extras = ["pytorch-lightning"], version = "^0.9.19"}
omegaconf = "^2.1.0"
matplotlib = "^3.4.2"
seaborn = "^0.11.1"
pytorch-lightning = "^1.3.7"
psutil = "^5.8.0"
sklearn = "^0.0"
hydra-core = "^1.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
